medallion-demo/requirements.txt
# Core
duckdb>=1.1.0
pyarrow>=17.0.0
deltalake>=0.22.0
pandas>=2.2.0
polars>=1.12.0

# Orchestration
apache-airflow>=2.10.0

# Data Quality
great-expectations>=1.2.0
pydantic>=2.9.0

# dbt
dbt-core>=1.9.0
dbt-duckdb>=1.9.0

# Testing
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
faker>=30.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.4.0
click>=8.1.0
pyyaml>=6.0.2
